renaming main_window_utils.py to treeview_utils.py
