+++ tests/ui_tests/test_main_window_utils.py
@@ -6,7 +6,7 @@ from PySide2.QtWidgets import QToolBar, QWidget, QTreeView
 
 from nexus_constructor.component_tree_model import ComponentTreeModel
 from nexus_constructor.component_tree_view import ComponentEditorDelegate
-from nexus_constructor.main_window_utils import (
+from nexus_constructor.treeview_utils import (
     create_and_add_toolbar_action,
     set_button_states,
     expand_transformation_list,
